requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0.1

# GUI framework (choose one based on preference)
# tkinter is included with Python, so no need to install
# For more advanced GUI:
# PyQt6>=6.5.0
# PySide6>=6.5.0

# Optional dependencies for advanced features
# rawpy>=0.18.0          # For RAW image processing
# matplotlib>=3.7.0      # For histograms and plots
# scipy>=1.11.0          # For advanced image processing algorithms